UV_Mapping_Pixel_ART/requirements.txt
Pillow==9.1.0
pygame==2.1.2
